move v2 alias tests
